# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "National Parks Finder API",
    "version": "1.0.0",
    "description": "The documentation for the National Parks Finder API"
  },
  "contact": {
    "email": "afbelardicode@gmail.com"
  },
  "host": "np-backend.herokuapp.com",
  "basePath": "/api",
  "schemes": ["https"],
  "tags": [
    {
      "name": "users",
      "description": "All info for each user"
    },
    {
      "name": "parks",
      "description": "Retrieve parks information"
    }
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "bearer",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer Token"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],


  "paths": {



    "/users/signup": {
      "post": {
        "tags": ["users"],
        "summary": "Sign up for an account",
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Issue creating account. Not successful"
          }
        },
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object that gets added to database to create an account",
            "required": true,
            "example": {
                "username": "exampleUser",
                "password": "examplePassword",
                "email": "exampleUser@example.com"
            }
          }
        ]
      }
    },



    "/users/login": {
        "post": {
          "tags": ["users"],
          "summary": "Log into user account",
          "responses": {
            "200": {
              "description": "Log in successful"
            },
            "400": {
              "description": "Issue logging in"
            }
          },
          "consumes": ["application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "User object for logging into account",
              "required": true,
              "example": {
                  "email": "exampleUser@example.com",
                  "password": "examplePassword"
              }
            }
          ]
        }
      },

      "/users/{userId}": {
        "get": {
          "tags": ["users"],
          "summary": "Obtain one user",
          "security": [{"Bearer": []}],
          "responses": {
            "200": {
              "description": "User successfully returned"
            },
            "400": {
              "description": "Issue getting user"
            }
          },
          "consumes": ["application/json"],
          "parameters": [
            {
              "in": "path",
              "name": "userId",
              "description": "User object for logging into account",
              "required": true,
              "type": "string"
            }
          ]
        }
      }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "favorites": [{
            "parkCode": {
                "type": "string"
            }
        }]
      }
    }
  }
}